testcase-eda-analysis-master/requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.18.8
botocore==1.21.8
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
colorama==0.4.4
contextvars==2.4
cryptography==3.4.7
dbiz-autotest-sdk==1.0.0
idna==2.10
immutables==0.16
importlib-metadata==4.8.1
iniconfig==1.1.1
jmespath==0.10.0
loguru==0.5.3
MouseInfo==0.1.3
numpy==1.19.5
packaging==21.0
pandas==1.1.4
paramiko==2.7.2
Pillow==8.3.2
pluggy==0.13.1
py==1.10.0
PyAutoGUI==0.9.53
pycparser==2.20
pyDes==2.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pytest==6.2.4
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
pywin32==301
PyYAML==5.3.1
requests==2.25.0
s3transfer==0.5.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
win32-setctime==1.0.3
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.5.0
pytest-dependency==0.5.1